Correction of work and refactoring

+++ AddExpression.cs
@@ -16,7 +16,7 @@ namespace ConsoleApp
             rightExpression = right;
         }
 
-        public int Interpret(Context context)
+        public double Interpret(Context context)
         {
             return leftExpression.Interpret(context) + rightExpression.Interpret(context);
         }
